[02] AI Agents/requirements.txt
openai
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
phidata
streamlit
phi
sqlalchemy
pgvector
psycopg[binary]
pypdf